Slice sample a binomial regression rate

Slice sample a binomial regression rate

mh_binom_reg(
  p,
  k,
  n,
  mean,
  precision,
  proposal = c("normal", "uniform", "quadratic taylor"),
  ...,
  proposal_sd = 1
)

ss_binom_reg(
  p,
  k,
  n,
  mean,
  precision,
  ...,
  w = 1,
  nexpand = 10,
  ncontract = 100
)

Arguments

p
the previous iteration of the logit-probability

k
the realized value from the binomial distribution

n
the number of trials

mean
the prior mean

precision
the prior precision

proposal
The method to use to propose new values. "normal" proposes normal deviates from the current value. "uniform" proposes uniform deviates. "quadratic taylor" proposes using a second-order Taylor approximation of the log-density at the current value, which amounts to a normal proposal with mean (usually) not equal to the current value.

...
Other arguments (not used)

proposal_sd
For "normal" proposals, the standard deviation(s) of proposals. For "uniform" proposals, the half-width of the uniform proposal interval.

w
The width of each expansion (to the right and left each)

nexpand
The maximum number of expansions (to the right and left each)

ncontract
The maximum number of contractions. If this is exceeded, the original value is returned

Details

ss_binom_reg slice samples and mh_binom_reg Metropolis-samples p conditional on k, n, mean, and precision, where k ~ Binom(n, expit(p)) and p ~ N(mean, precision).

In the case that n is zero, slice sampling is ignored in favor of a normal draw. In the special case when an entire (multivariate) row of n is zero, the entire row is simultaneously drawn (in R); when only some elements are zero, they're drawn univariately (in C++).

Both vectorized over p, k, n, and mean. If precision is a matrix, p is assumed to be multivariately distributed, and different internals are used.

The internals are defined in C++.

See also