Slice sample a multinomial regression rate

Slice sample a multinomial regression rate

ss_multinom_reg(
  p,
  z,
  k,
  mean,
  precision,
  ref = c("first", "last"),
  ...,
  w = 1,
  nexpand = 10,
  ncontract = 100
)

Arguments

p
the previous iteration of the logit-probability, as an array of dimension r x i x j, where r is the number of realizations, i is the number of (correlated) betas (that is, the dimension of precision), and j is the number of classes to split between. It is expected that p[, , ref] == 0

z
a matrix of zeros and ones. The zeros determine so-called "structural zeros": outcomes which are not possible. This is assumed not to change over the first dimension (r) of p. This is of dimension i x j.

k
the realized value from the binomial distribution; the same size as p (r x i x j).

mean
the prior mean for p. Must be either a scalar or an array of size r x i x (j-1)

precision
an array of dimension i x i x (j-1). The j-dimension is assumed to be independent.

ref
One of "first" or "last", denoting which column of p is the reference. If "first", then mean and precision map to the second through j-th elements of p. If "last", then mean and precision map to the first through (j-1)-th elements of p.

...
Other arguments (not used)

w
The width of each expansion (to the right and left each)

nexpand
The maximum number of expansions (to the right and left each)

ncontract
The maximum number of contractions. If this is exceeded, the original value is returned

Details

The internals are defined in C++.

See also