Slice sample a Poisson regression rate

Slice sample a Poisson regression rate

ss_pois_reg(L, k, mean, precision, ..., w = 1, nexpand = 10, ncontract = 100)

Arguments

L
the previous iteration of the log-rate

k
the realized value from the Poisson distribution

mean
the prior mean

precision
the prior precision

...
Other arguments (not used)

w
The width of each expansion (to the right and left each)

nexpand
The maximum number of expansions (to the right and left each)

ncontract
The maximum number of contractions. If this is exceeded, the original value is returned

Details

This function slice samples L conditional on k, mean, and precision, where k ~ Pois(exp(L)) and L ~ N(mean, precision).

In the case that k is NA (akin to ss_binom_reg in the case n == k == 0), slice sampling is ignored in favor of a normal draw. In the special case when an entire (multivariate) row of k is NA, the entire row is simultaneously drawn (in R); when only some elements are NA, they're drawn univariately (in C++).

This is vectorized over L, k, and mean. If precision is a matrix, L is assumed to be multivariately distributed, and a different function is used.

The internals are defined in C++.

See also