Other Utilities

Other Utilities

chol_inv(x)

chol_mvrnorm(
  n = 1,
  mu,
  Sigma,
  Precision,
  A = if (missing(Sigma)) backsolve(chol(Precision), diag(1, nrow(Precision))) else
    t(chol(Sigma))
)

logit(p)

expit(x)

Arguments

x
A matrix for chol_inv or a numeric vector for expit

n
Number of samples to draw

mu, Sigma, Precision
parameters for a multivariate normal distribution. One of Sigma and Precision must be supplied.

A
the "square root" by which we multiply the independent normals. Used only for speed.

p
A numeric vector

Details

Unlike MASS::mvrnorm(), chol_mvrnorm() uses the Cholesky decomposition to draw random samples. This is usually faster; the largest gains are seen when n is small (when n is large, the matrix multiplication becomes the limiting factor).